% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMITAB.R
\name{cleanMITAB}
\alias{cleanMITAB}
\title{\code{cleanMITAB} extracts interactor Uniprot (or else, specified by interactor_IDs_databases) IDs, interactor taxonomy IDs, Publication Identifiers, Confidence values and generates unique pair ID}
\usage{
cleanMITAB(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.5 or 2.7 formats. Details: \code{\link{queryPSICQUIC}}}
}
\value{
data.table for MITAB 2.5: containing the interactor Uniprot IDs, interactor taxonomy IDs, Publication Identifiers, Confidence values and unique pair ID(alphanumerically sorted)
}
\description{
\code{cleanMITAB} extracts interactor Uniprot (or else, specified by interactor_IDs_databases) IDs, interactor taxonomy IDs, Publication Identifiers, Confidence values and generates unique pair ID
}
\details{
Output column description:

pair_id - unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs

IDs_interactor_A, IDs_interactor_B - interacting molecule ID

interactor_IDs_databases_A, interactor_IDs_databases_B - database that provides interacting molecule ID such as UniProt, ChEMBL or IntAct (IntAct: when the molecule cannot be mapped to ID in any other resource)

Taxid_interactor_A, Taxid_interactor_B - taxonomic species that interacting molecule belongs to

Publication_Identifiers - pubmed ID of the papers that has reported the interaction

Confidence_values - MIscore. Details: \link{https://psicquic.github.io/MITAB25Format.html}, \link{http://www.ebi.ac.uk/intact/pages/faq/faq.xhtml}, \link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4316181/}
}
\author{
Vitalii Kleshchevnikov
}
