% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gcross.R
\name{Gcross}
\alias{Gcross}
\title{Multitype Nearest Neighbour Distance Function (i-to-j)}
\usage{
Gcross(X, i, j, r = NULL, correction = "km")
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of the cross type distance distribution function \eqn{G_{ij}(r)} will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor).}

\item{i}{The type (mark value) of the points in \code{X} from which distances are measured. A character string (or something that will be converted to a character string).}

\item{j}{The type (mark value) of the points in \code{X} to which distances are measured. A character string (or something that will be converted to a character string).}

\item{r}{numeric vector. The values of the argument \eqn{r} at which the distribution function \eqn{G_{ij}(r)} should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument.}

\item{correction}{Character string specifying the edge correction(s) to be used. Options are \code{"none", "rs", "km", "hanisch"} and \code{"best"}. Alternatively \code{correction="all"} selects all options.}
}
\value{
An object of class "\code{fv}" in \pkg{spatstat}.

Essentially a data frame containing columns
\item{r}{the values of the argument \eqn{r} at which the function \eqn{G_{ij}(r)} has been estimated}
\item{raw}{the uncorrected estimate of \eqn{G_{ij}(r)}, i.e., the empirical distribution of the distances from each point of type \eqn{i} to the nearest point of type \eqn{j}}
\item{theo}{the theoretical value \eqn{G_{ij}(r)=r} for a stationary Poisson process}
together with columns named \code{"rs"}, \code{"km"}, and/or \code{"han"}, according to the selected edge corrections. These columns contain estimates of the function \eqn{G_{ij}(r)} obtained by the edge corrections named. If \code{correction = "km"}, an additional column \code{hazard} provides the hazard rate \eqn{\lambda(r)} of \eqn{G_{ij}(r)} by the spatial Kaplan-Meier method.
}
\description{
For a multitype point pattern, estimate the distribution of the distance from a point of type \eqn{i} to the nearest point of type \eqn{j}.
}
\examples{
# load pre-processed lung cancer data
load(url("https://github.com/julia-wrobel/MI_tutorial/raw/main/Data/lung.RDA"))

# randomly select a lung cancer image and define cell types
library(dplyr)
subj_df = filter(lung_df, image_id == "#73 6-333-711_[44841,18646].im3")
subj_df = subj_df \%>\%
   mutate(phenotype = case_when(
      phenotype_cd19 == "CD19+" ~ "CD19+ B-cell",
      phenotype_cd14 == "CD14+" ~ "CD14+ cell",
      phenotype_cd8 == "CD8+" ~ "CD8+ T-cell",
      phenotype_cd4 == "CD4+" ~ "CD4+ T-cell",
      phenotype_ck  == "CK+" ~ "CK+ cancer cell",
      TRUE ~ "Other"
   ))

# convert the selected lung cancer image into the object of class "ppp" representing a multitype point pattern
pp_lung = multitype.pp(subj_df$x, subj_df$y, subj_df$phenotype)

# calculate estimates of the G-cross function between CD19+ B-cell and CD14+ cell
G = Gcross(pp_lung, "CD19+ B-cell", "CD14+ cell", correction = "all")
plot(G, main = "G-cross")
}
