% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gest.R
\name{Gest}
\alias{Gest}
\title{Nearest Neighbour Distance Function G}
\usage{
Gest(X, i, r = NULL, correction = "rs")
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of \eqn{G(r)} will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor).}

\item{i}{The type (mark value) of the points in \code{X} from which an estimate of\eqn{G(r)} will be computed. A character string (or something that will be converted to a character string).}

\item{r}{Optional. Numeric vector. The values of the argument \eqn{r} at which \eqn{G(r)} should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument.}

\item{correction}{Optional. The edge correction(s) to be used to estimate \eqn{G(r)}. A vector of character strings selected from \code{"none", "rs", "km", "Hanisch"} and \code{"best"}. Alternatively \code{correction="all"} selects all options.}
}
\value{
An object of class "\code{fv}" in \pkg{spatstat}.

Essentially a data frame containing some or all of the following columns:
\item{r}{the values of the argument \eqn{r} at which the function \eqn{G(r)} has been estimated}
\item{rs}{the “reduced sample” or “border correction” estimator of \eqn{G(r)}}
\item{km}{the spatial Kaplan-Meier estimator of \eqn{G(r)}}
\item{hazard}{the hazard rate \eqn{λ(r)} of \eqn{G(r)} by the spatial Kaplan-Meier method}
\item{han}{the Hanisch correction estimator of \eqn{G(r)}}
\item{theo}{the theoretical value of \eqn{G(r)}for a stationary Poisson process of the same estimated intensity.}
}
\description{
Estimates the nearest neighbour distance distribution function \eqn{G(r)} from a point pattern in a window of arbitrary shape.
}
