% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lest.R
\name{Lest}
\alias{Lest}
\title{L-function}
\usage{
Lest(X, i, r = NULL, correction = "isotropic")
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of \eqn{L(r)} will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor).}

\item{i}{The type (mark value) of the points in \code{X} from which an estimate of \eqn{L(r)} will be computed. A character string (or something that will be converted to a character string).}

\item{r}{Optional. Vector of values for the argument \eqn{r} at which \eqn{L(r)} should be evaluated. Users are advised not to specify this argument; there is a sensible default.}

\item{correction}{Optional. A character vector containing any selection of the options \code{"none", "border", "bord.modif", "isotropic", "Ripley", "translate", "translation", "rigid", "periodic", "good"} or \code{"best"}. It specifies the edge correction(s) to be applied. Alternatively \code{correction="all"} selects all options.}
}
\value{
An object of class "\code{fv}" in \pkg{spatstat}.

Essentially a data frame containing columns
\item{r}{the vector of values of the argument \eqn{r} at which the function \eqn{L} has been estimated}
\item{theo}{the theoretical value \eqn{L(r) = r} for a stationary Poisson process}
together with columns named \code{"border", "bord.modif", "iso"} and/or \code{"trans"}, according to the selected edge corrections. These columns contain estimates of the function \eqn{L(r)} obtained by the edge corrections named.
}
\description{

}
