% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitype.pp.R
\name{multitype.pp}
\alias{multitype.pp}
\title{Create a Multitype Point Pattern}
\usage{
multitype.pp(
  x,
  y,
  marks,
  xrange = c(min(x), max(x)),
  yrange = c(min(y), max(y)),
  poly = NULL,
  unitname = NULL
)
}
\arguments{
\item{x}{Vector of \eqn{x} coordinates of data points.}

\item{y}{Vector of \eqn{y} coordinates of data points.}

\item{marks}{mark values for the points. A vector or data frame.}

\item{xrange}{\eqn{x} coordinate limits of enclosing box}

\item{yrange}{\eqn{y} coordinate limits of enclosing box}

\item{poly}{Optional. Polygonal boundary of window. Incompatible}

\item{unitname}{Optional. Name of unit of length. Either a single character string, or a vector of two character strings giving the singular and plural forms, respectively.}
}
\value{
An object of class "ppp" describing a multittype point pattern in the two-dimensional plane.
}
\description{
Creates an object of class "ppp" representing a multitype point pattern dataset in the two-dimensional plane.
}
