% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcfcross.R
\name{pcfcross}
\alias{pcfcross}
\title{Multitype pair correlation function (cross-type)}
\usage{
pcfcross(X, i, j, r = NULL, correction = c("isotropic", "Ripley", "translate"))
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of the cross-type pair correlation function \eqn{g_{ij}(r)} will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor).}

\item{i}{The type (mark value) of the points in \code{X} from which distances are measured. A character string (or something that will be converted to a character string).}

\item{j}{The type (mark value) of the points in \code{X} to which distances are measured. A character string (or something that will be converted to a character string).}

\item{r}{Vector of values for the argument \eqn{r} at which \eqn{g_{ij}(r)} should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument.}

\item{correction}{Character string specifying the edge correction(s) to be used. Options are \code{"isotropic", "Ripley", "translate"} and \code{"best"}. Alternatively \code{correction="all"} selects all options.}
}
\value{
An object of class "\code{fv}" in \pkg{spatstat}.

Essentially a data frame containing columns
\item{r}{the vector of values of the argument \eqn{r} at which the function \eqn{g_{ij}(r)} has been estimated}
\item{theo}{the theoretical value \eqn{g_{ij}(r) = 1} for independent marks}
together with columns named \code{"iso"} and/or \code{"trans"}, according to the selected edge corrections. These columns contain estimates of the function \eqn{g_{ij}(r)} obtained by the edge corrections named.
}
\description{
Calculates an estimate of the cross-type pair correlation function for a multitype point pattern.
}
