% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomInteractome.R
\name{randomInteractome}
\alias{randomInteractome}
\title{Retrieve molecular interactions for the random set of proteins (of a particular taxon)}
\usage{
randomInteractome(MITABdata = NULL, degree_data = NULL, n_inter,
  degree = NULL, ...)
}
\arguments{
\item{MITABdata}{data.table containing pre-loaded molecular interaction data as returned by \code{\link{queryPSICQUICrlib}}, usefull for taking multiple samples, the default in NULL}

\item{degree_data}{data.table containing pre-calculated (using \code{\link{edgelist2degree}}) degree for each node in MITABdata, usefull for taking multiple samples, the default in NULL}

\item{n_inter}{integer (1L), the number of proteins for which to retrieve the random set of interactions}

\item{degree}{integer (\code{n_inter} L), specify the degree for \code{n_inter} number of proteins to produce the network with the specific degree distribution, if set to NULL (default) the degree distribution will correspond to that of \code{taxid} interactome}
}
\description{
Retrieve molecular interactions for the random set of proteins (of a particular taxon)
}
\details{
Random network can be specified to have specific degree distribution. If the (\code{degree} parameter is set \code{taxid} proteins will be split by degree and from each degree group a sample of the size specified by how many times specific degree number is repeated in \code{degree} will be taken.

If the degree distribution is not specified a sample of \code{n_inter} is taken from all proteins which have interaction data available in the \code{database} for \code{taxid}. In this case, the degree distribution of the resulting set of proteins will be similar to the degree distribution in the interactome of \code{taxid} in \code{database}.

\code{randomInteractome} retrieves molecular interactions using \code{\link{fullInteractome}}
}
\examples{
# retrive the interactome using PSICQIUC servise (or by reading local copy) from IMEx databases for a list of 200 random human (9606) proteins, not specifying their degree distribution
set.seed(1)
random = randomInteractome(n_inter = 200, degree = NULL, taxid = "9606", database = "imex", protein_only = TRUE)
# retrive the interactome from MITABdata for a list of 200 random human (9606) proteins, not specifying their degree distribution
full = fullInteractome(taxid = "9606", database = "imex", format = "tab25", clean = TRUE, protein_only = TRUE)
set.seed(1)
random = randomInteractome(MITABdata = full, n_inter = 200, degree = NULL)
}
\author{
Vitalii Kleshchevnikov
}
